# docs/openapi/catalog-api-v1.json
{
  "components": {
    "parameters": {
      "RequestId": {
        "description": "Optional client correlation ID. The service generates and returns one when omitted.",
        "in": "header",
        "name": "X-Request-ID",
        "required": false,
        "schema": {
          "maxLength": 128,
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "CertificationRequestCreate": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "CertificationRequestCreate",
        "type": "object"
      },
...
      "CertificationRequestRead": {
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decision_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision By"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/UsageDecisionStatus"
          }
        },
        "required": [
          "id",
          "asset_id",
          "requested_by",
          "status",
          "decision_by",
          "decision_note",
          "requested_at",
          "decided_at"
        ],
        "title": "CertificationRequestRead",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UsageDecisionStatus": {
        "enum": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "UsageDecisionStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT or OIDC access token with a tenant claim and one or more DataGenie roles.",
        "scheme": "bearer",
        "type": "http"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "DataGenie API Support",
      "url": "https://github.com/itismohan/datagenie/issues"
    },
    "description": "## DataGenie Catalog API\n\nThe DataGenie Catalog API provides tenant-isolated metadata discovery, stewardship,\ngovernance, operational lineage integration boundaries, quality-aware search, and\ncustomer operations. Every endpoint is versioned under `/api/v1`.\n\n### Authentication and tenant boundary\n\nSupply an OAuth/OIDC or HS256 bearer token in `Authorization: Bearer <token>`.\nOutside local development, every token must include the configured tenant claim.\nResponses, search facets, audit history, exports, retention policies, and webhook\noperations are restricted to that tenant.\n\n### Request correlation and errors\n\nClients may send `X-Request-ID`; otherwise the service generates one and returns it\non every response. Error responses use a stable envelope containing an application\n`code`, a human-readable `message`, and the request ID required for support.\n\n### Idempotency and rate limits\n\nUse `Idempotency-Key` for supported mutating operations, especially source and\ningestion-job creation. Catalog endpoints return standard `RateLimit-*` and\n`Retry-After` headers when request protection is active.\n",
    "license": {
      "name": "Proprietary"
    },
    "summary": "Tenant-isolated metadata governance and discovery platform",
    "title": "DataGenie Catalog API",
    "version": "1.0.0",
...
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/governance/assets/{asset_id}/certification-requests": {
      "post": {
        "operationId": "request_certification_api_v1_governance_assets__asset_id__certification_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificationRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificationRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "A valid bearer token is required.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A valid bearer token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "The authenticated principal is not authorized for this operation.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The authenticated principal is not authorized for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "code": "rate_limit_exceeded",
                        "message": "Too many requests. Retry after the supplied interval.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many requests. Retry after the supplied interval."
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "internal_error": {
                    "value": {
                      "error": {
                        "code": "internal_error",
                        "message": "An unexpected server error occurred.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Request Certification",
        "tags": [
          "Governance"
        ]
      }
    },
    "/api/v1/governance/assets/{asset_id}/classification-detections": {
      "post": {
        "operationId": "detect_sensitive_classification_api_v1_governance_assets__asset_id__classification_detections_post",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClassificationFindingRead"
                  },
                  "title": "Response Detect Sensitive Classification Api V1 Governance Assets  Asset Id  Classification Detections Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "A valid bearer token is required.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A valid bearer token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "The authenticated principal is not authorized for this operation.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The authenticated principal is not authorized for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "code": "rate_limit_exceeded",
                        "message": "Too many requests. Retry after the supplied interval.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many requests. Retry after the supplied interval."
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "internal_error": {
                    "value": {
                      "error": {
                        "code": "internal_error",
                        "message": "An unexpected server error occurred.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Detect Sensitive Classification",
        "tags": [
          "Governance"
        ]
      }
    },
    "/api/v1/governance/assets/{asset_id}/classification-findings": {
      "get": {
        "operationId": "get_classification_findings_api_v1_governance_assets__asset_id__classification_findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClassificationFindingRead"
                  },
                  "title": "Response Get Classification Findings Api V1 Governance Assets  Asset Id  Classification Findings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "A valid bearer token is required.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A valid bearer token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "The authenticated principal is not authorized for this operation.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The authenticated principal is not authorized for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "code": "rate_limit_exceeded",
                        "message": "Too many requests. Retry after the supplied interval.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many requests. Retry after the supplied interval."
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "internal_error": {
                    "value": {
                      "error": {
                        "code": "internal_error",
                        "message": "An unexpected server error occurred.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Classification Findings",
        "tags": [
          "Governance"
        ]
      }
    },
    "/api/v1/governance/assets/{asset_id}/quality-evidence": {
      "put": {
        "operationId": "put_asset_quality_evidence_api_v1_governance_assets__asset_id__quality_evidence_put",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetQualityEvidenceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "A valid bearer token is required.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A valid bearer token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "The authenticated principal is not authorized for this operation.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The authenticated principal is not authorized for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "rate_limit_exceeded": {
                    "value": {
                      "error": {
                        "code": "rate_limit_exceeded",
                        "message": "Too many requests. Retry after the supplied interval.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many requests. Retry after the supplied interval."
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "internal_error": {
                    "value": {
                      "error": {
                        "code": "internal_error",
                        "message": "An unexpected server error occurred.",
                        "request_id": "b9ce11cc-4e68-4f4c-9e0f-4c1f0e2c6e2d"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Put Asset Quality Evidence",
        "tags": [
          "Governance"
        ]
      }
    },
...
  },
  "servers": [
...
}